python/requirements.txt
Flask
jsonschema
fastapi
uvicorn
numpy
Jinja2
PyYAML
